requirements-dev.txt
# requirements-dev.txt
pytest
python-dotenv
